requirements.txt
APScheduler==3.10.4
config==0.5.1
dependencies==7.7.1
fastapi==0.115.14
fastapi_users==14.0.1
passlib==1.7.4
pydantic==2.11.7
python-dotenv==1.1.1
python_jose==3.5.0
Requests==2.32.4
SQLAlchemy==2.0.41
sqlmodel==0.0.24
